% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addlmeq.R
\name{addlmeq}
\alias{addlmeq}
\title{Add Linear Fitting Equation}
\usage{
addlmeq(fit, decimal = 2)
}
\arguments{
\item{fit}{an object of class "lm" produced by simple linear models.}

\item{decimal}{displayed decimal places.}
}
\value{
An equation expression for the given simple linear fitting.
}
\description{
\code{addlmeq} produces an equation expression for a simple linear fitting to be added in a plot.
}
\examples{
fit = lm(dist ~ speed, data = cars)
plot(cars)
abline(fit, col = 'red')
title(addlmeq(fit, 2), adj = 0.1, line = -1.8, col.main = 'blue', cex.main = 1.1)
}
\seealso{
\code{\link{addr2}}.
}
