% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addr2.R
\name{addr2}
\alias{addr2}
\title{Add R-Square}
\usage{
addr2(obs, pred, decimal = 2)
}
\arguments{
\item{obs}{vector of observation.}

\item{pred}{vector of prediction.}

\item{decimal}{displayed decimal places.}
}
\value{
An r2 expression for the given vectors of observation and prediction.
}
\description{
\code{addr2} produces an r-squared (r2) expression (r2 = ...) given a vector of observation and prediction.
}
\examples{
fit = lm(dist ~ speed, data = cars)
plot(cars)
abline(fit, col = 'red')
title(addlmeq(fit, 2), adj = 0.1, line = -1.8, col.main = 'blue', cex.main = 1.1)
title(addr2(cars$dist, predict(fit)), adj = 0.08, line = -2.8, col.main = 'blue', cex.main = 1.1)
}
