% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Bias}
\usage{
bias(obs, pred)
}
\arguments{
\item{obs}{vector of observation.}

\item{pred}{vector of prediction.}
}
\value{
A data.frame containing r2, bias and RMSE.
}
\description{
\code{bias} computes r-squared (r2), bias and root mean square error (RMSE) given a vector of observation and prediction.
}
\examples{
bias(1:10, 1:10 + rnorm(10))
}
