% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbat.R
\name{gbat}
\alias{gbat}
\title{Grouped Boxplot at}
\usage{
gbat(n.group, n.subgroup, gap1 = 2, gap2 = 1)
}
\arguments{
\item{n.group}{number of groups.}

\item{n.subgroup}{number of sub-groups.}

\item{gap1}{group gap.}

\item{gap2}{sub-group gap.}
}
\value{
A list of complete x-axis locations and each group's center location.
}
\description{
\code{gbat} generates grouped boxplot locations.
}
\examples{
gbat(5, 2, 2, 1)

}
\seealso{
\code{\link{gboxplot}}.
}
