% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gboxplot.R
\name{gboxplot}
\alias{gboxplot}
\title{Grouped Boxplot}
\usage{
gboxplot(
  f,
  data,
  col = c("green", "red"),
  alpha = 1,
  at1 = NULL,
  at2 = NULL,
  at1.labels = NULL,
  at2.labels = NULL,
  at1.centered = FALSE,
  mgp1 = c(3, 0.2, 0),
  mgp2 = c(3, 0.25, 0),
  lwd.axis = 0.1,
  legend = FALSE,
  legend.position = "topright",
  legend.label1 = "Sub-group 1",
  legend.label2 = "Sub-group 2",
  cex.legend = 1,
  x.intersp = 1,
  y.intersp = 1,
  inset = c(0, 0),
  ...
)
}
\arguments{
\item{f}{a formula such as \eqn{y ~ g1:g2}, where y is a numeric vector of data values to be split into groups. Note that \eqn{~ g1 + g2} is equivalent to \eqn{g1:g2}.}

\item{data}{a data.frame or data.table from which the variables in formula should be taken.}

\item{col}{box colors.}

\item{alpha}{color transparency (a number between \code{0} and \code{1}).}

\item{at1}{if null (default), even-distributed locations would be generated for the boxes. Otherwise, users should provide a "at list" generated from \code{\link{gbat}} to adjust the positions where the boxes should be drawn.}

\item{at2}{the points at which y-tick-marks are to be drawn.}

\item{at1.labels}{x-tick-labels. Valid only when at1.centered = \code{TRUE}. Otherwise please use \code{names} to change x-tick-labels.}

\item{at2.labels}{y-tick-labels.}

\item{at1.centered}{logical. If \code{TRUE} x-axis ticks would be centered.}

\item{mgp1}{mgp for x axis, same as the mgp in \code{\link{par}}.}

\item{mgp2}{mgp for y axis, same as the mgp in \code{\link{par}}.}

\item{lwd.axis}{axis line width.}

\item{legend}{logical. Whether a legend should be added.}

\item{legend.position}{position of legend.}

\item{legend.label1}{legend label for sub-group 1.}

\item{legend.label2}{legend label for sub-group 2.}

\item{cex.legend}{size of legend.}

\item{x.intersp}{horizontal spacing between legend item and text.}

\item{y.intersp}{vertical spacing between legend items.}

\item{inset}{inset distance from the margins as a fraction of the plot region when legend is placed by keyword.}

\item{...}{other graphical parameters to be passed.}
}
\value{
List with components same as \code{\link{boxplot}}.
}
\description{
\code{gboxplot} produces grouped box plot.
}
\examples{
opar = par(no.readonly = TRUE)
par(cex.axis = 0.9, mai = c(0.5, 0.4, 0.3, 0.2), mfrow = c(2, 2), oma = c(1, 5, 2, 5))

# base r boxplot ----
boxplot(
  len ~ supp:dose, data = ToothGrowth,
  boxwex = 0.5, col = c('red', 'green'),
  xlab = '', ylab = 'tooth length',
  ylim = c(0, 35), yaxs = 'i',
  sep = ':',
  main = 'base r boxplot'
)

# gboxplot with defaul at and legend ----
gboxplot(
  len ~ supp:dose, data = ToothGrowth,
  boxwex = 0.5, col = c('red', 'green'), alpha = 0.5,
  xlab = '', ylab = 'tooth length',
  ylim = c(0, 35), yaxs = 'i',
  sep = ':',
  main = 'gboxplot with defaul at and legend',
  legend = TRUE, legend.position = 'bottomright', legend.label1 = 'OJ', legend.label2 = 'VC', cex.legend = 0.8,
  x.intersp = 0.8, y.intersp = 0.8
)

# gboxplot with adjusted at ----
n.group = 3
n.subgroup = 2
# adjusted box positions
myat = gbat(n.group, n.subgroup, 1.2, 0.35)
gboxplot(
len ~ supp:dose, data = ToothGrowth,
boxwex = 0.2, col = c('red', 'green'), alpha = 0.5,
xlab = '', ylab = 'tooth length',
at1 = myat,
ylim = c(0, 35), yaxs = 'i',
sep = ':',
main = 'gboxplot with adjusted at',
legend = TRUE, legend.position = 'bottomright', legend.label1 = 'OJ', legend.label2 = 'VC', cex.legend = 0.8,
x.intersp = 0.8, y.intersp = 0.8
)

# gboxplot with adjusted at and centered ticks ----
gboxplot(
  len ~ supp:dose, data = ToothGrowth,
  boxwex = 0.2, col = c('red', 'green'), alpha = 0.5,
  xlab = '', ylab = 'tooth length',
  at1 = myat, at1.labels = c(0.5, 1, 2), at1.centered = TRUE,
  ylim = c(0, 35), yaxs = 'i',
  sep = ':',
  main = 'gboxplot with adjusted at and centered ticks',
  legend = TRUE, legend.position = 'bottomright', legend.label1 = 'OJ', legend.label2 = 'VC', cex.legend = 0.8,
  x.intersp = 0.8, y.intersp = 0.8
)
par(opar)

}
\seealso{
\code{\link{gbat}}.
}
