% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splot.R
\name{splot}
\alias{splot}
\title{X-Y Plotting}
\usage{
splot(
  data = NULL,
  axis.style = "r",
  sec.x = F,
  sec.y = F,
  at1 = NULL,
  at2 = NULL,
  at3 = NULL,
  at4 = NULL,
  at1.labels = NULL,
  at2.labels = NULL,
  at3.labels = NULL,
  at4.labels = NULL,
  mgp1 = c(3, 0.1, 0),
  mgp2 = c(3, 0.2, 0),
  lwd.axis = 0.1,
  lwd.tick = 0.5,
  col.axis = "black",
  bg = NULL,
  ...
)
}
\arguments{
\item{data}{a two-column \emph{data.frame} or \emph{data.table} with x variable in the first column and y variable in the second.}

\item{axis.style}{axis style, can be one of "i" (default) or "r".}

\item{sec.x}{logical. If TRUE the axis on the top would be added.}

\item{sec.y}{logical. If TRUE the axis on the right would be added.}

\item{at1}{the points at which x-tick-marks are to be drawn.}

\item{at2}{the points at which y-tick-marks are to be drawn.}

\item{at3}{the points at which sec.x-tick-marks are to be drawn.}

\item{at4}{the points at which sec.y-tick-marks are to be drawn.}

\item{at1.labels}{x-tick-labels.}

\item{at2.labels}{y-tick-labels.}

\item{at3.labels}{sec.x-tick-labels.}

\item{at4.labels}{sec.y-tick-labels.}

\item{mgp1}{mgp for x axex, same as mgp() in par() in base r.}

\item{mgp2}{mgp for y axex, same as mgp() in par() in base r.}

\item{lwd.axis}{lwd.axis axis line width.}

\item{lwd.tick}{lwd.tick tick line width.}

\item{col.axis}{axis color.}

\item{bg}{background color.}

\item{...}{other graphical parameters to be passed.}
}
\description{
\code{splot} handels generic 2-d plotting.
}
\examples{
splot(cars)

}
